requirements.txt
Django==2.0
django-haystack==2.8.1
elasticsearch==2.4.1
pdfdocument==4.0.0
pdfminer==20191125
Pillow==7.1.2
pkg-resources==0.0.0
pycryptodome==3.9.7
pytz==2019.3
reportlab==3.5.42
urllib3==1.25.9
